requirements.txt
torch>=2.2
torchvision>=0.17
timm>=0.9
opencv-python>=4.9
numpy>=1.26
matplotlib>=3.8
scikit-image>=0.22
scikit-learn>=1.4
xgboost>=2.0
mediapipe>=0.10
piexif>=1.1
exifread>=3.0
# jpegio>=0.2 ; platform_system != "Windows"
reportlab>=4.0
pydantic>=2.7
fastapi>=0.111
uvicorn>=0.29
streamlit>=1.36
pillow>=10.3
pyyaml>=6.0
pytest>=8.2
black>=24.8
ruff>=0.6
numpy>=2,<2.3
